[TASK] Replace classname strings with ::class constants
This replaces the use of class names in strings with using the ::class
constant instead. This has the advantage of immediately being able to
spot errors (the class will be marked as undefined in any proper IDE)
and when refactoring the classes can be identified precisely. Also it
makes the use if namespace imports possible.

Change-Id: I85a60ea1469b683ade306e13afd71e3d80f7bdbb
Releases: master
Reviewed-on: http://review.typo3.org/40894
Reviewed-by: Karsten Dambekalns <karsten@typo3.org>
Tested-by: Karsten Dambekalns <karsten@typo3.org>
Reviewed-by: Bastian Waidelich <bastian@typo3.org>
Reviewed-by: Christian Müller <christian.mueller@typo3.org>
Tested-by: Christian Müller <christian.mueller@typo3.org>
Original-Commit-Hash: 91e59edadfbf7e28bc02083b315ed6d2bd11f827